WL#11722: Step 23 ----------------- Rewrote the scanCheck function. This function performs the end of range check. It has the end boundary that is static for the scan operation and compares this to each key that is to be reported back. The compare operation previously used functions in NdbPack.cpp that built up a lot of context again and again.
The new compare function is very simple. It builds up an array of
objects that have two parts:
1) m_data_ptr: Pointing to the column value
2) m_data_len: The length of the column value
NULL is represented by length equal to 0.
For the end of range boundary this is calculated by first unacpking the
boundary and then using the NdbPack::Desc and NdbPack::BoundC objects
to create a new NdbPack::BoundArray. The above array of representations
of the bound object is placed in a new NdbPack::DataArray object.

The key is read into an array with Attrinfo setup. The Attrinfo is
kept and a NdbPack::DataArray is created that points to data in
the Attrinfo object. This method is very quick and only consumes
20-30 instructions for a simple boundary with 1 column in the
bound condition. This mapping is done in the new function
NdbPack::DataArray::init_poai (poai stands for plain old attribut
info).

In addition the readKeyAttrsCurr function was dropped and scanCheck
calls tuxReadAttrsCurr directly from scanCheck.

This change improved performance by 25% in a scan filtering
benchmark.