WL#11722: Step 26 ----------------- The interface to read and update columns in DBTUP ships two 32-bit words that describe the column being updated. The first is shipped as an argument and the second is shipped as a variable in the KeyReqStruct.
Nowadays most CPUs are 64-bit CPUs, so it makes sense to pass
both of those through as one argument on the stack. CPUs can
execute shift operations very quickly so the overhead of
packing them up and down is small. This decreases the risk of
cache misses and waiting for loads to have their data accessible.