Bug#24734190 KILL-SERVER : LOG PARSING BUFFER OVERFLOW. RECOVERY MAY HAVE FAILED!
Problem
--------
While doing a crash recovery, the server failed as it ran out of space on the recovery parsing buffer. This was because the server received a log record which was larger than
the recovery parsing buffer which is initialized to 2MB.

Solution
---------
Allocate the recovery parsing buffer dynamically and increment it until it reaches the value set by the parameter log_buffer_size.

Reviewed by: Debarun Banerjee (debarun.banerjee@oracle.com)

RB: 19211