BUG#20383215 - POSIX_FALLOCATE() CAN BE INTERRUPTED (OS ERROR NUMBER 4)
Because posix_fallocate() could be interrupted, the related error messages
may be printed out in this case, which makes the internal MTR test case
failing with errors/warnings. Since this error message about the
interruption may not help users too much, it is proper to just ignore
this error of EINTR and go through to write zeroes to the file.
Along with this change, it now also tries writing zeroes when the error
is EINVAL.

RB: 19056
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>