Bug#27604471 THERE IS NO LOCAL IP ADDRESS MATCHING THE ONE CONFIGURED FOR THE LOCAL NODE
Note
====
This bug is a duplicate of Bug#27684295 BACKPORT BUG#27376511 TO 5.7.

Problem
=======
When using an IP or hostname in the GR configuration parameters, there would
be no match on Mac OSX. This occurs because the buffer isn't large enough
to store all the interfaces' configurations, and its size does not increase
because the buffer isn't completely filled, since the configurations have
variable sizes which do not add up exactly to the buffer size.

Fix
===
Modified condition to increase the buffer when the difference between its
size and that of the retrieved configurations is smaller than IFNAMSIZ
added to the size of the sockaddr_storage struct, which is "large enough
to accommodate all supported protocol-specific address structures", according
to IETF RFC3493.

Reviewed-by: Sujatha Sivakumar <sujatha.sivakumar@oracle.com>
Reviewed-by: Filipe Campos <filipe.campos@oracle.com>
RB: 19146