BUG 27672154 - SET ROLE DEFAULT CRASHES UNDER SKIP-GRANT-TABLES
Description
------------

In case of --skip-grant-tables none of the global  acl objects are initialized
so does the g_default_roles object.
In this case we try to access the uninitialized the object hence crash.

Fix
---
1. Added a nullptr check.
2. If no default roles are found then list is set to empty

Tesing
------
Added the scenario in the skip_grants.test file

Itch run --
push id: 12759126
2018-03-19 10:39:58

Review
-------
RB#19187