Bug#27454299 SYSTEM USERS USING MYSQL_NATIVE_PASSWORD WHEN              CACHING_SHA2_PASSWORD IS DEFAULT
Description
------------
While fixing this bug last time I chose the 1000 iterations to generate the
digest for these 3 system users. I assumed that they are locked users and not
used for internal purpose.

Harin however had a different opinion that we must be consistent with the
default behavior of caching_sha2_passwprd plugin which does the 5000 iterations
to create digests. It will be consistent and there is not really any performance
impact during initilzation/upgrade. Hence this change.

Review:
-------
RB#18957