Bug#27704268 MOVE DICTIONARY LOCKS OUTSIDE THE PARSER
The parser should not take any dictionary locks, so this patch moves
the acquisition of dictionary locks regarding SRID values on a column
from the parsing stage to the execution stage.

An added bonus of this moving is that the code added for bug#27220467
in order to release all the locks properly is not needed anymore.
This is due to the fact that at during the execution stage, there is
already a machinery in place to release all the lock taken even in
the event of an error.

Change-Id: Ie4608576660a95ec75e576c7bafc5c75d399e42f