Bug #28094147: PREPARATORY CHANGES FOR WL #12074 [remove QEP_TAB from init]
Remove the QEP_TAB parameter from Init; give it to the constructor instead
where it makes sense, or better yet, just give in the parameters from it that
we need.

Having a forced QEP_TAB in Init() no longer makes sense when we get iterators
that no longer correspond to a single table.

Change-Id: Idd5d1afb068aa90cbabdeaf1e7cbf3889ca89e97