Bug#27756083 "INSERT INTO SELECT" QUERY INSERTS GEOMETRY OF DIFFERENT TYPE
When inserting data into a geometry column, the server tries to ensure
that the data inserted is a valid geometry. Since we also have specific
geometry types like POINT, LINESTRING etc, the server tries to ensure
that the type of the inserted geometry is correct. However, in the case
of "INSERT INTO tbl SELECT", the server managed to omit many of the checks
that are in place to ensure the correct geometry type. This patch makes
the server do the same checks for "INSERT INTO tbl SELECT" as it does
for "INSERT INTO tbl VALUES (...)".

While fixing this, it was discovered that we have some test for this
specific (mis)behavior. These tests were removed since they trigger an
unwanted behavior. The new tests that are added covers up for the
removed tests.

Change-Id: Ib2057a654d439a2e19943fd789c1a03d57971a9f