Bug#27769521 WRONG ERROR WHEN CREATE TABLE FAILS IN NDB [noclose]
Faulty error message is returned when CREATE TABLE fails to create
the table or other object in NDB. For example the error
"Table 'x' dosen't exist" may be returned. This patch
adds a tetscase showing the faulty behaviour.

Preconditions:
 1) Max length of a table name in NDB is currently 128 bytes.
 2) The MySQL Server creates tables in NDB using name on the form
     <db_name>/<catalog_name>/<table_name> where <catalog_name> is
     always hardcoded to "def".
 3) The current limit of identifier names in MySQL is 64 characters,
    but for NDB tables the limit is currently only 63 bytes
    due to a legacy bug.

Using two maximum length identifiers(i.e 63 bytes) it's possible to
hit the 128 bytes limit in NDB. Unfortunately the ndbcluster plugin
does not detect this error correctly and returns the error:
  "Table '<long identifier>' doesn't exist"

Change-Id: I0db6dc6a200061be654e7dd9360a8e1b3abbe87c