BUG#27481872 HANDLE_FATAL_SIGNAL (SIG=11) IN TRACE::FUNCTION_ENTER ON INSTALL PLUGIN
Problem:
When innodb is in read only mode, for instance when
innodb-force-recovery is set to 6, installing the semi-sync master plugin would
result in unexpected errors.

Analysis:
With the introduction of data dictionary, the plugin initialization
order changed between 5.7 and 8.0: first the plugin registers itself in the
plugin table and only then it runs its specific initialization function.
However, if an error occurs at this stage, the cleanup assumes that the
plugin-specific initialization function was executed and tries to
execute the plugin-specific deinitialization function.

Fix:
If the semisync master plugin was not initialized, do not use ack_receiver
if it is empty.