Bug#27788925: LOG_ERROR_SERVICES COMPONENTS SHOULD BE COMMA-SEPARATED
Global system variable log_error_services accepts a list of services.
Services could previously be separated by ';' (semicolon).
This is still the case, but ',' (comma) may be used instead if preferred.
Using both separators in the same list is not allowed.
A separator may not occur before the first list-element (that is, before
the first service).
All whitespace padding is ignored.