Bug #27542720   SCHEMA MISMATCH - TABLE FLAGS DON'T MATCH, BUT FLAGS ARE NUMBERS
PROBLEM

During a schema mismatch of flags when tablespace import is attempted
the concerned server table flags are not printed instead number of
columns are printed , also the table flags are not human readable.

FIX

Print the table flags instead of columns and in human readable string
format for the user's convinience.

Reviewed by: Allen Lai<zheng.lai@oracle.com>
RB :19156