WL#9855 - MySQL GCS: Instrument memory usage in GCS/XCom
This worklog instruments relevant memory management operations in GCS/XCom
and exposes them automatically in P_S tables metrics, in order to provide
users with useful information for monitoring and debugging. As a result of
this worklog, the source code of GCS/XCom has been modified to use the
Server's Performance Schema Interface to register keys to be associated with
the instrumented memory elements and update the memory usage statistics in
P_S tables.

In particular, this worklog instruments memory operations that occur within
the XCom cache. As a result, users will have access to information regarding
the amount of entries in the cache (both current and aggregated) as well as
the size of the cached data (current and overall).