Bug #27627136 DOXYERROR.LOG DOESN'T CONTAIN ERRORS FROM PLANTUML AND DIA
Post push fix to filter out error messages that occur when not building
from a git repo.

Doxygen runs a filter that tries to run a few git commands. These fail
when the source directory is not a git repository, and the error
messages show up in the doxyerror.log file.

Change-Id: I07d3e67dc405367b91b83c47bbd35e0c4d5247bb