Fix different issues with the ATR test testOIBasic: Patch is an effort to try to hunt down random AutoTest failures in this test where it looks like duplicates of test rows appears from time to time. Item 1) below is a possible root cause for these failures.
Test framework depended on a set of unique characters to be
randomly generated during test setup. The collation of the character
set to be used is also randomly picked among those available.

Problems fixed:

1) When checking that a generated character was unique, it was memcmp'ed
against all the other character. However a character collation may
define two characters to be equal even if not binary equal. Any
'collation insensitive' (*_ci) collations will for instance compare
upoper and lower case characters to be equal. Thus we had to replace
memcmp by using the collation specific strnncollsp() function.

2) Furthermore the test framework organized the gennerated characters
in an ordered list, doing a strnxfrm and memcmp'ing the result.
Simplified this by using the strnncollsp() function for compare.

3) Need for doing strnxfrm() became obsolete due to 2). Removed usage
and related datastructures for holding the xfrm'ed result.

4) Simplified logic to randomly construct a 'well_formed' random character.

5) Fixed a valgrind detected problem of using delete to free 'tablist'
created with new[]