Bug #25873310   FULLTEXT SEARCH CAN NOT FIND WORD WHICH CONTAINS "," OR "."
PROBLEM

Ngram fts parser used to parse "," and "." as words and were tokenized
which is wrong and uneccessary.

FIX

Disallow "." and "," as words to be tokenized (as is the case with
native parser) and that would solve the bug issue which provides
different results in case of natural / boolean search.

Reviewed by: Jimmy Yang<Jimmy.Yang@oracle.com>
RB: 19191