WL#9131 Add RETRIES field to replication_applier_status_by_worker P_S table
This worklog introduces eight new columns in the performance schema
table replication_applier_status_by_worker that report the number of
retries for the last transaction applied by the worker and the
transaction the worker is currently applying, and the number, message
and timestamp of the last transient error causing a retry.
The worklog also changes the behaviour of the column
APPLYING_TRANSACTION_START_APPLY_TIMESTAMP which, instead of being reset
every time the transaction was retried, stores the same timestamp when
the worker first started applying it.