Bug#27300658: IN MTS SQL THREAD EXITING AND INITIALIZING MESSAGE WITH DIFFERENT POSITION
Problem:
========
When MTS is used and slave executes STOP SLAVE and START
SLAVE, SQL thread reports different position for exiting and
initializing position.

Analysis:
=========
When SQL thread stops either due to regular STOP SLAVE
command or due to an error it prints the master log file
name and pos upto which replication is consistent. In case
of MTS the reported file name and position represent the low
water mark upto which replication stream is consistent and
there are no gaps.

But as reported in the bug scenario the LWM which is
reported is not the most recent LWM. The reason behind this
scenario is, SQL thread first reports the current LWM
specific positions and then invokes 'slave_stop_workers'
function call to stop the worker threads. During worker
threads stop 'mts_checkpoint_routine' is called which will
update the LWM and hence the SQL thread's master log file
name and position get advanced when compared to the file
name and position that were printed in error log.

Fix:
===
Print the SQL thread positions after the
'slave_stop_workers' call which invokes MTS checkpoint.
This will ensure that the applier positions are in sync with
the latest LWM.