Bug#27830283 DROP SCHEMA PERFORMANCE_SCHEMA IS ALLOWED
Before this fix, the statement
  DROP SCHEMA performance_schema;
was allowed, resulting in a corrupted installation.

The root cause is historical,
as database creation scripts before MySQL 8.0 used
to DROP + CREATE the performance_schema schema
during database installation with an external SQL script,
so there was a valid use case for DROP SCHEMA.

Starting with 8.0, initialization is done within the server,
so that allowing DROP SCHEMA (to a user) is no longer necessary.

The fix is to strengthen ACL checks for the performance_schema,
in PFS_internal_schema_access::check().