Bug#27842735: Sql_cmd::get_owner() returns stale object after repreparation
Prepared_statement::reprepare() prepares a temporary
Prepared_statement object that is allocated on the stack. During its
preparation, the temporary object sets itself as the owner of the
Sql_cmd object in the LEX. Since the temporary Prepared_statement
object is automatically destroyed when reprepare() returns, the
Sql_cmd object ends up with a dangling owner pointer.

This does not cause any problems at the moment, since
Sql_cmd::get_owner() is currently only called during preparation,
while the Prepared_statement object pointed to is still alive. It
might be a source of future bugs if more calls to get_owner() are
added other places in the code.

The fix is to make Prepared_statement::swap_prepared_statement() set
the owner back to itself before it returns.