Bug#26523254 - MODE == 16 || !FIX_BLOCK->PAGE.FILE_PAGE_WAS_FREED FOR DROP TABLE
This is due to some complex scenario:
1. Some ALTER TABLE ... ADD INDEX tries to create a new index I on table T1
2. However, it fails afterwards, so I is marked as aborted. Note that the
related dict_index_t is not removed from memory, because some other thread
is accessing this table. So dict_index_t of I is only marked as
ONLINE_INDEX_ABORTED_DROPPED.
3. In this post_ddl, the physical index was freed, and the
file_page_was_freed for the root page is now set to false
4. Since the root page has been freed, another CREATE INDEX on T2 may take
over this page, but this time it's not a root page.
5. Then this new CREATE INDEX may free this page, because this is the last
page of one level. This happens in BtrBulk::finish(). At this stage,
the page is set to free explicitly, with file_page_was_freed set to true.
6. Now a command to drop T1 is issued. Since this is a table residing in
a general tablespace, it has to drop each index one by one. So DDL logs
to free index are written, including freeing index I.
7. In the post_ddl of this DROP TABLE, it will check the status of I,
and find that the root page of I is now marked as file_page_was_freed=true,
which is not expected, thus the assertion.

This basically would not make any further bug, just checking the flag fails. However, it indeed makes no sense to free a already freed index in step 7, so in step 6, it's not necessary to write DDL log for this index I, if the status of this index is not ONLINE_INDEX_COMPLETE, supposing any failed index should be dropped previously.

RB: 19377
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>