Bug 27777959 INNODB: ASSERTION FAILURE: UT0UT.CC DUE TO DEADLOCK AT FSP_PAGE_CREATE()
Problem:

In a single thread, two mtrs are latching the first page of LOB.  One of that mtr
is freeing the first page of LOB and commits.  But the other mtr is still holding
the latch.  Another thread is allocating the page and it gets this page, resulting
in a deadlock.

Solution:

Ensure that only one mtr is latching the first page of LOB.