Bug#27851644: Compiler warning (maybe-uninitialized) in mysql_create_user()
GCC 7 emits the following warning if the -Og option is used:

sql/auth/sql_user.cc: In function ‘bool mysql_create_user(THD*, List<LEX_USER>&, bool, bool)’:
sql/auth/sql_user.cc:1796:21: error: ‘acl_user’ may be used uninitialized in this function [-Werror=maybe-uninitialized]
           grant_role(acl_role, acl_user, false);
           ~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~

This patch reorganizes the code so that it is easier for the compiler
(and hopefully also for humans) to see that variable is always
initialized.

The call to grant_role() was previously guarded by a check for the
condition (result == 0 && not_granted). The call is moved into the
preceding if statement, where both of these conditions are known to
hold. This allows narrowing the scope of the maybe-uninitialized
variable (and a few others) so that it can be initialized when it is
declared.