Bug #27858762: OUT-OF-BOUNDS ACCESS IN M_RECORD_POINTERS
When we use LIMIT with priority queue, the record pointers are already
set up before we start reading records, so we should not reset the buffer.
Doing so would cause us to read and write records past the end of the
vector, which is undefined behavior (but happened to work fine in practice,
since the memory was not freed).

Change-Id: If4f0a059c403021071ed3d0ff07f6e2d8113204c