Bug #26236298 : ADD --DEFAULTS-FILE AND FAMILY TO NDB_TOP
Add support of options --defaults-file and family [--print-defaults,
--no-defaults, --defaults-extra-file, --defaults-group-suffix] from group
[client] of a configuration file.
Add --socket option, with short option -S, to allow both socket and port
connection, with higher preference to socket.

Note that the --login-path option is already supported.

Note : The previous commit [7bfa251] was for version 7.6 while this commit is
for versions 8.0 and above with the additional parameter MEM_ROOT in the
defaults file option required for these versions.