Bug#27849973 ABORT THE DEBUG SERVER IN CASE OF ASSERTIONS IN THE STD LIBRARY ON WINDOWS
Change implementation of my_parameter_handler, which is used as an
argument to _set_invalid_parameter_handler on Windows. This will abort the
server if the C/C++ runtime reports "out of range" or similar errors.

Fix failing mtr test 'main.import' on windows, reading an empty file
is a no-op.

Make the error-handling of a failing create_pid_file() deterministic on windows.

Change-Id: Ib81c614636d6791ba2f7afd191bbdac0ae8a0d57