WL #11785: Volcano iterator design [patch 6/13, CALC_FOUND_ROWS deoptimization]
Remove a very narrow special-casing of CALC_FOUND_ROWS that would allow it to
skip scanning a MyISAM table (instead picking out the actual number of rows
directly from the statistics, or from the filesort input) under specific conditions.

This simplifies the next sorting patch, as filesort no longer needs to save the
number of scanned rows in TABLE (the sort result, which was the natural place to
save it before, is being moved out of TABLE).

Change-Id: Ia9fa7d15b818e565269911de41b2691ea1bbd27c