WL#9638: Step 1 --------------- To understand the use of disk storage by Partial LCPs over the time of several LCPs with updates, deletes and inserts and combinations of those we have written an LCP simulator tool. This tool is not normally compiled, it is very easy to compile. Compile it with the following command:
gcc lcp_simulator.c
This produces an a.out file that can be executed. If a different
simulation is desired one can easily change the source code
in set_rates to decide on the rates of insert, update and delete
for each LCP id, the number of LCPs that are simulated can be
set in the program using the constant NUM_SIMULATED_LCPS.

The size of LCPs when using only updates can be fairly easily
modelled using mathematics. But modelling inserts and deletes
behaviour mathematically is a bit harder, so this simulation
tool gives lots of input into how Partial LCPs behave.