WL#9638: Step 2 --------------- Based on the analysis from the LCP simulator it is clear that during heavy insert activity we can decrease the amount of LCP writes, similarly we have to increase it during heavy delete activities.
Since heavy insert is an activity that puts a lot of strain to the
computer, it is good to be able to lower the LCP writes during the
insert activity. Lowering these writes during heavy insert means that
the LCP size on disk will temporarily grow, e.g. we only record the new
records in LCPs during inserts, the LCP size on disk will temporarily
grow to 150% overhead whereas if we write at 40% of the speed of update
activity we will only have around 65% overhead. The extra overhead of
decreasing the LCP write speed to 40% is less than 5%.

To handle this we introduce a new configuration variable
InsertRecoveryWork, this defaults to 40%. It can be set as low as 0%
and can be set at most to 70%. Setting it to higher than 70% has no
benefits.

We increase RecoveryWork default from 50% to 60% to decrease the
LCP write sizes to have a better chance to survive heavy insert
activities.

This is integrated into the current Partial LCP algorithm and we estimate
the amount of insert and delete activity through keeping track of the
row count at the previous LCP. We only care for the change in number of
rows, if there has been inserts and deletes that take out each other
these will be treated as updates.