Bug#27755892 - CRASH IN ROW_SEL_FIELD_STORE_IN_MYSQL_FORMAT_FUNC
This is a regression from this commit: ef20b9. The cause is that with
the changes in aforementioned commit, optimizer supports 'using index'
for some LIKE queries on a prefix index, including the virtual index.
Before this change, this kind of query will go with table scan.
So the assertion in row_sel_field_store_in_mysql_format_func() is
somehow bogus now. So the fix is to screen out the case that the field
is a virtual column. In case of a virtual column, it just ignores the
field prefix length checking.

RB: 19455
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>