WL#9638: Step 7 --------------- Introducing two levels of bitmaps to filter rows quicker for LCPs. The highest level is 8 bits, each bit is set if a row within a 4096 bytes area is updated. The first word of the row is used to decide which part the row is counted in. The lower level is 128 bits, each bit is set if the row within a 256 bytes area is updated.
If none of the 8 bits on highest level is set we will skip the
page entirely.

When the LCP_DELETE or LCP_SKIP bit is set on a row it is important
to set the bit again to ensure that the row is handled also in the
next LCP.

A proof is provided that this method actually provides the desired
result for recovery.

Prefetching have not been fixed in this patch and some extra debug
checks are not introduced yet.