BUG#27664539 INNODB: ASSERTION FAILURE: BUF0FLU.CC:457:              (BUF_POOL->FLUSH_LIST).START == __NULL
If no-redo mini transaction modifies a given page first time, but the
page was stored on disk before, it must ensure that the lsn set to the
newest_modification of the page is not smaller than the page_lsn stored
on disk (available at frame + FIL_PAGE_LSN).

Added assertion that ensures that the new page_lsn is never smaller than
the previous one. It is checked whenever page is flushed to disk, when
the newest_modification is being copied to field at frame + FIL_PAGE_LSN.