BUG#90526 [MySQL 8.0 GA Release Build] Assertion failure: write_max_size > 0 BUG#90523 [MySQL 8.0 GA Release Build] InnoDB Assertion: (capacity & (capacity - 1)) == 0 BUG#90522 [MySQL 8.0 GA Release Build] InnoDB: Assertion failure: (n & (n - 1)) == 0
We misinterpreted what PLUGIN_VAR_NOSYSVAR was supposed to provide.
It was documented as "Not a server variable" which seemed to suggest
that the variable is hidden (especially as it was then being skipped
when setting values to sysvars).

We tested and observed that sysvars could neither set via option file
nor during runtime. But we missed that the variable still could be set
via command line option.

In result the new EXPERIMENTAL sysvars were configurable via cmd line,
even when ENABLE_EXPERIMENT_SYSVARS was not configured for cmake.

Restored the previous name - NOSYSVAR and implemented EXPERIMENTAL
sysvars in different way - wrapped their definition by a simple ifdef.

This is the most safe way to do it. In future we could try to think
about introducing PLUGIN_VAR_EXPERIMENTAL to make life easier - as
we had to set default values for all underlying global variables,
because we can no longer rely on sysvars initialization to default.