BUG#27899678: CHECK-TESTCASE FAILS IF SERVER IS STARTED WITH               --DISABLED_STORAGE_ENGINES=INNODB
Description:
------------
MTR's check-testcase run fails if a server is started with
--disabled_storage_engines=InnoDB option.

Issue:
------
check-testcase creates a temporary table which is used to check if all
the temporary files created within a test are cleaned up at the end.
However, the create temporary table statement will fail if a test
starts the server with 'disabled_storage_engines=InnoDB' option.

Fix:
----
Check if CREATE TEMPORARY TABLE query is executed successfully, if it
fails with error ER_DISABLED_STORAGE_ENGINE, skip the check for
cleanup of temporary files created withing a test.

Change-Id: Id7b230ec7ff5b1e45ca0982354b94f37ce96627c