Bug#27858055	BACKPORT --EXEC_IN_BACKGROUND MTR COMMAND TO 5.7
Problem: Currently in MTR, there is no way to execute a shell command
in the background.
For eg: executing MYSQLSLAP should be done using --exec but it is not going
into background even if appened "&" at the end of the --exec command.

Fix: Implementing a new command --exec_in_background that puts
the shell command in the background.
USAGE: --exec_in_background $MYSQL_SLAP --create-schema=test --delimiter=";"
                            --iterations=500 --query="INSERT INTO t1 values (1)"
                            --concurrency=10 --silent 2>&1