WL#9638: Step 16 ---------------- Added more debug handling for LCP_SKIP handling and also for page filtering.
Fixed a bug where the page filtering could exit a page in a
new place, this new place also needed to handle setting the
D value for the page.

Fixed a bug in refactoring of the code where the label
found_tuple was called without proper initialisation of the
thbits variable.