Bug #27646160: Fix FK retrieval issue in COPY_FK_FOR_OFFLINE_ALTER
In ha_ndbcluster::copy_fk_for_offline_alter(), if the Alter_drop list
has a foreign key marked for dropping, its name is retrieved and is
checked in the Ndb Dictionary if such a FK actually exists. To do that,
list of all dependent objects on the table is obtained and then the list
is searched for any available FK with the name mentioned in the drop
list.

But when searching for the foreign keys in the dependent list, the
elements are not being filtered based on the ForeignKey. Due to this,
there is a chance that the function might think of an index or a trigger
as a foreign key item and end up trying to retreive it and ultimately
fail. This patch fixes that by skipping further search if the element in
the dependent list is not a foreign key.