Bug#27439587: Add/Drop FK over schema distribution doesn't reload referenced tables' metadata
When a foreign key is added or dropped as a part of any DDL, all the
parent tables referenced by the foreign key should be flushed out of the
server cache to force reload the FK metadata in the handler. This is
already done by the fix for Bug#24666177. And also the cached parent
tables in the server Data Dictionary should be reloaded. Server does
this as part of all DDLs. But when this DDL is distributed to other
mysqlds connected to the same cluster, through schema disribution, the
same invalidations are not being done to the parent tables' caches in
that mysqld server. Due to this, any subsequent queries relying on the
foreign key metadata information of the corresponding parent tables
might return inconsistent results.

Fix for versions 8.0+:
During schema distribution, in the participant mysqld, the older
definition can be retrieved from its Data Dictionary. And any new
foreign key definitions can be retrieved from the NDB Dictionary. A
collection of referenced tables from these new and old foreign key
definitions can be extracted and they are locked before actually
executing the distributed DDL operation. After executing the DDL, the
referenced tables are invalidated from the Data Dictionary and also
flushed out from the server.

A new class - Ndb_referenced_tables_invalidator is added to handle all
this operations. It has methods for fetching and invalidating the
referenced tables.

A new header file - ndb_fk_util.h is added as a part of this patch. This
file along with its cpp implementation has couple of utility function
definitions required by the Ndb_referenced_tables_invalidator class.