Bug #27903792 [MYSQL 8.0 GA DEBUG BUILD] ASSERTION `DATA_TYPE() == MYSQL_TYPE_GEOMETRY' FAILED
Problem: Debug builds fail an assertion during UNION queries when
computing the combined data type of a GEOMETRY column and SELECT * FROM
(SELECT NULL).

Item_type_holder::join_types() calls geom_type() on items without
checking that the items are of type MYSQL_TYPE_GEOMETRY. In this case
the item is an Item_field with data type MYSQL_TYPE_GEOMETRY, and
Item_field::geom_type() fails an assertion.

Solution: Set the combined geometry type to be Field::GEOM_GEOMETRY if
the item is not of a geometry type.

Change-Id: Ib16438be6d754bacd3aae92f6730e92068f0b157