BUG#26553164 - INNODB PARTITION TABLE HAS UNEXPECTED ROW LOCK
This is similar issue to bug#26731025, the only difference is that
this one is for ordered index scan, while the other one is for
unordered index scan. The cause is same too, that is for the equal
range scan, it used to call improper read_range* function and set
an incorrect lock to the record, which leads to the final lock waiting.
The fix is that for equal range scan, call the proper
index_next_same_in_part() to set a proper record to previous locking
waiting.

RB: 19528
Reviewed-by: Debarun Banerjee (debarun.banerjee@oracle.com)