WL #12074: Volcano iterator executor base [4/15, joins]
Support inner and outer nested loop joins as iterators. Unlike in the old
executor, nested loop joins are not special-cased in any way (they're just
a regular iterator), which means that adding e.g. hash join as a first-class
citizen is straightforward, provided that the optimizer can plan them.

We need two new additions to the RowIterator interface: First, a new member
function called SetNullRowFlag() that asks the iterator to mark its current
row buffers as NULL rows or not, used for outer joins. Second, some auxiliary
functions for turning on and off performance schema batch mode, intended to
reduce the PFS overhead when scanning the innermost table.

Note that this fixes a few corner cases around line numbers in error messages.

Change-Id: Ie5c9ab85fa7456196f35c6da4778b5f30f5ce020