Bug #27938507: IMPROVE LINE NUMBER REPORTING IN WARNINGS
Change the reported “line numbers” (really row numbers) on warnings
in SELECT statements from being based on number of input rows read
to the actual row number the user sees. In particular, if joining
three tables one-to-one and there's a warning on the second row of
the first table, the warning should say “at line 2”, not “at line 4”.
(Note that the latter depends on the join order chosen!)
Similarly for GROUP BY: If aggregating e.g. ten and ten rows,
and there's a warning when processing an aggregate in the second
group, we should say “at line 2”, not “at line 19” or similar.

There are still issues with this notion, particularly if we need to
emit a warning for a row that is later filtered away (we'll say
e.g. “at line 123” when the warning is for a row that would have
been at line 123 if it were not filtered away). However, the new behavior
is strictly less confusing than the old one, even in this case.

Change-Id: I200974857bf866a18c50265d4d7b52c34c0b1c18