Bug#26450129:DEAD BRANCH IN SEARCH_KEY_IN_TABLE()
Problem:

if (key_type & UNIQUE_KEY_FLAG && table->s->uniques)
  {
    ...
  }
The condition (table->s->uniques) is always 0 so the if block will never be
executed. In cases where different primary key is used on master and slave
and if there is an Unique key present on slave then that should be used for
lookup created index. Before this bug in above cases a TABLE_SCAN was being
performed which can be slower than INDEX_SCAN that will be performed after
the bug fix.

Fix:
Remove the condition (table->s->uniques).