BUG#27898284 TRX->DECLARED_TO_BE_INSIDE_INNODB
Problem :
--------
The root cause is we are allowing to set value beyond the limits of
innodb-concurrency-tickets. The documented limits are as follows.

innodb_concurrency_tickets
==========================
Type 	         integer
Minimum Value 	 1
Maximum Value 	 4294967295

When we try to set it to a value beyond integer limit 4G, we are holding
it in long variable and letter it gets truncated while type casting to
integer value. When the value is set to [4294967295 + 1 = 4294967296]
it gets truncated to 0 which is invalid.

The zero value is unexpected for innodb_concurrency_tickets and results in
assert later.

Solution :
----------
Set the upper limit for innodb_concurrency_tickets to 4294967295
as per documentation. When the value is set beyond, it is adjusted
to 4294967295 and a truncation warning is thrown. This is inline
with the behaviour of other configurations.

Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>

RB: 19546